refactor: composed reusable func and better error message
